extract_glade: add linenos, context, GtkBuilder
This commit adds support for source line numbers, context attrs in
source XML, and formally adds support for the new GtkBuilder (GTK3
and Glade 3.0+) "UI Definition XML" format. The older glade2 format
is not supported in the latest versions of Glade Interface Designer,
so it's probably time to support it properly.

New features and code changes:

* Report line numbers for the XML formats.
* Extract "context" attrs to "msgctxt" POT fields.
* Make the extraction respect Babel "keywords" (i18n function names).
* Add docstring for extract_glade().
* Document that we're ignoring comment tags.
* Update the test cases.
* Update setup.py, but do not bump version.